requirements.txt
asttokens==3.0.0
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
executing==2.2.0
fonttools==4.56.0
h5py==3.13.0
ipython==8.32.0
jedi==0.19.2
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.10.0
matplotlib-inline==0.1.7
numpy==2.2.3
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
stack-data==0.6.3
threadpoolctl==3.5.0
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
eli5==0.13.0
graphviz==0.20.3
shap==0.46.0 
kaggle==1.6.17
xgboost==2.1.4
pytorch
fastai